test(service): split field-validation tests along create/update boundary
`service_field_validation_tests.rs` grew to 510 lines, tripping the
`linecheck` 500-line-per-file gate. Split it into
`service_field_validation_create_tests.rs` (svc_create_* tests, plus the
shared TempHome/make_routine/create_req_with_title helpers) and
`service_field_validation_update_tests.rs` (svc_update_* tests, which
reach the shared helpers via `super::service_field_validation_create_tests`).
Test-only reorganization, no behavior change.
